EVM-IT: Break up fevm.rs
