+++ testing/integration/tests/evm/features/SelfDestruct.feature
@@ -2,9 +2,21 @@ Feature: SelfDestruct
 
   # Some of these are impossible to test today because we don’t dispatch to smart contract logic on send.
   # https://github.com/filecoin-project/ref-fvm/issues/835
+  # To simulate these, we will explicity call a method on the contract to trigger the call back to the self-destructed contract.
 
   Scenario: SELFDESTRUCT on contract creation, sending funds to self => fails
+    Given 1 random account
+    When the beneficiary is set to 0x0000000000000000000000000000000000000000
+    Then account 1 fails to create a SelfDestructOnCreate contract
+
+
   Scenario: SELFDESTRUCT on contract creation, sending funds to an f410 EthAccount that doesn’t exist => succeeds
+    Given 1 random account
+    And a non-existing f410 account 0x76c499be8821b5b9860144d292fff728611bfd1a
+    When the beneficiary is set to 0x76c499be8821b5b9860144d292fff728611bfd1a
+    And account 1 creates a SelfDestructOnCreate contract
+    Then the f410 account 0x76c499be8821b5b9860144d292fff728611bfd1a exists
+
 
   @dispatch_on_send
   Scenario: SELFDESTRUCT on contract creation, sending funds to a smart contract that ends up returning the funds to the sender => in Eth those funds would vanish, in Filecoin they are preserved.
